+++ relvar-core/src/query/tests/mod.rs
@@ -1,2 +1,2 @@
-pub mod basic;
-pub mod common;
+pub(crate) mod basic;
+pub(crate) mod common;
